Final code for now!
Massively updated the code. Might continue modifying the structure, but for now it's much clearer to read and the interactions between functions make quite a bite more sense (particularly the simulation/how the automata is run). Code is also more logically broken up. This is the real starting point for the repo